% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritized_grouping.R
\name{file_extension}
\alias{file_extension}
\title{Title}
\usage{
file_extension(filenames)
}
\arguments{
\item{filenames}{character vector of file name. Splits at the last '.'.}
}
\value{
character vector of extension
}
\description{
Title
}
\examples{
file_extension("data/prioritized_sample.csv")
}
