% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formats.R
\name{parse_prio_form}
\alias{parse_prio_form}
\title{Parse input data with columns of priorities to columns of groups}
\usage{
parse_prio_form(data, id = 1, prio.cols, sort.cols = FALSE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{id}{id column. Numeric index or column name. Default is 1.}

\item{prio.cols}{priority columns. Numeric indices or column names.}

\item{sort.cols}{flag to sort priority columns names/indices. Default=FALSE}
}
\value{
data.frame
}
\description{
This handles transforming data from a typical Google form to the format
compatible with \code{prioritized_grouping()}.
}
