% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formats.R
\name{parse_string_form}
\alias{parse_string_form}
\title{Parse input data from column of strings with prioritised group names}
\usage{
parse_string_form(data, id = 1, string.col, pattern = NULL)
}
\arguments{
\item{data}{data.frame or tibble}

\item{id}{id column. Numeric index of column name. Default is 1.}

\item{string.col}{string column. Numeric index or column name.}

\item{pattern}{regex pattern to use for splitting priorities string with
\code{strsplit()}.
Default is ";".}
}
\value{
data.frame
}
\description{
This handles transforming data from a typical Microsoft form to the format
compatible with \code{prioritized_grouping()}.
}
