% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritized_grouping.R
\name{plot.prioritized_groups_list}
\alias{plot.prioritized_groups_list}
\title{Plot extension for easy groupings plot}
\usage{
\method{plot}{prioritized_groups_list}(data, ...)
}
\arguments{
\item{data}{data of class 'prioritized_groups_list'}

\item{...}{passed on to 'grouping_plot()'}
}
\value{
ggplot2 list object
}
\description{
Plot extension for easy groupings plot
}
\examples{
read.csv(here::here("data/prioritized_sample.csv")) |>
  prioritized_grouping() |>
  plot(overall = TRUE, viridis.option="D",viridis.direction=-1)
}
